Set size to amount of space required
Fix sequencer expands to fill the space its in and pushing other
elements off below it.

The sequencer doesn't need to fill the space and it's likely that you
want to draw additional information about the selection below it.
Instead of filling available space, fill horizontal space and calculate
how much space we need.