Fixed potential div0 floating point issue
if you application has float point exceptions activated, you get a div0 when
calculating stepLength, even if the value isn't really used in the loop.

By moving it into the loop we prevent this error - at the cost of doing
the division multiple times. It's imo still better than relying on the fact
that float point exceptions must be disabled on the user side to use ImGuizmo.