Fix EditTransform() example in README
The previous example code in the README would not compile with latest
ImGui. This commit fixes the compile and makes it more closely align
with the example code.

Update ImGui::IsKeyPressed() to match latest API
Update ImGui::InputFloat3() to match latest API
Add ID to ImGui::Checkbox() call (not having this triggers an assert)
Add default: case to avoid warning
Change arguments to be more generic & align with example code